Created QuestionFactory and moved DataModels from ViewController
