+++ MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,0 +1,15 @@
+//
+//  QuizStepViewModel.swift
+//  MovieQuiz
+//
+//  Created by Юрий Демиденко on 22.08.2022.
+//
+
+import UIKit
+
+// модель для состояния "Вопрос задан"
+struct QuizStepViewModel {
+    let image: UIImage?
+    let question: String
+    let questionNumber: String
+}
