+++ MovieQuiz/Services/QuestionFactoryProtocol.swift
@@ -1,0 +1,12 @@
+//
+//  QuestionFactoryProtocol.swift
+//  MovieQuiz
+//
+//  Created by Юрий Демиденко on 22.08.2022.
+//
+
+import UIKit
+
+protocol QuestionFactoryProtocol {
+    func requestNextQuestion() -> QuizQuestion?
+}
